chore(one-to-many): move action to account folder
